requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.0.0
anthropic>=0.3.0
tweepy==4.14.0
schedule==1.2.0
python-dotenv==1.0.0
jinja2==3.1.2
lxml>=4.6.0
newspaper3k==0.2.8
python-dateutil==2.8.2
pydantic>=2.0.0
loguru==0.7.2
Flask==3.0.0
yagmail==0.15.293
psycopg2-binary==2.9.7
gunicorn==21.2.0
